Large scale refactor Changing the id structure to ulid for better aesthetics in the frontend

+++ src/main/java/com/divjazz/recommendic/global/general/ResponseWithCount.java
@@ -2,5 +2,5 @@ package com.divjazz.recommendic.global.general;
 
 import java.util.Set;
 
-public record ResponseWithCount<T>(Set<T> elements, int total) {
+public record ResponseWithCount<T>(Set<T> elements, long total) {
 }
